backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
channels==4.3.2
channels-redis==4.3.0
Pillow==12.0.0
PyMySQL==1.1.2
requests==2.32.3
beautifulsoup4==4.12.3
tenacity==9.0.0
tqdm==4.67.1
playwright==1.48.0
